Split `system::signal`
